+++ src/coreclr/scripts/cse_ml/jitml/__init__.py
@@ -1,8 +1,8 @@
 """JIT Machine Learning (JITML) is a Python library for the .Net JIT's reinforcement learning algorithms."""
-from.constants import MethodContext, CseCandidate, JitType
+from .method_context import MethodContext, CseCandidate, JitType
 from .superpmi import SuperPmi
 from .jit_cse import JitCseEnv
-from .machine_learning import JitRLModel
+from .machine_learning import JitCseModel
 from .conversions import get_observation
 
-__all__ = ['SuperPmi', 'JitCseEnv', 'JitRLModel', 'get_observation', 'MethodContext', 'CseCandidate', 'JitType']
+__all__ = ['SuperPmi', 'JitCseEnv', 'JitCseModel', 'get_observation', 'MethodContext', 'CseCandidate', 'JitType']
